fix: correcting received types from api;

+++ app/src/main/java/com/example/weatherapp/data/data_source/api/data_source/current_weather_data_source/LocationDataSource.kt
@@ -4,11 +4,11 @@ import com.google.gson.annotations.SerializedName
 
 data class LocationDataSource(
     val country: String,
-    val lat: Int,
+    val lat: Double,
     val localtime: String,
     @SerializedName("localtime_epoch")
-    val localtimeEpoch: Int,
-    val lon: Int,
+    val localtimeEpoch: Double,
+    val lon: Double,
     val name: String,
     val region: String,
     @SerializedName("tz_id")
